+++ src/examples/sequence-identity/index.ts
@@ -3,7 +3,7 @@ import {RcsbFv3DSequenceIdentity} from "../../RcsbFv3D/RcsbFv3DSequenceIdentity"
 
 document.addEventListener("DOMContentLoaded", function(event) {
 
-    const groupId: string = "584_30";
+    const groupId: string = "494_30";
     const panel3d = new RcsbFv3DSequenceIdentity({
         config:{
             groupId,